reorganize i18n and other tenant specific utils into views/tenants/utils dir
