+++ src/components/charts/new_revocations/RevocationsByGender/helpers.js
@@ -20,7 +20,7 @@ import pipe from "lodash/fp/pipe";
 import set from "lodash/fp/set";
 import toInteger from "lodash/fp/toInteger";
 import { calculateRate } from "../helpers/rate";
-import { translate } from "../../../../views/tenant_copy_translations/i18nSettings";
+import { translate } from "../../../../views/tenants/utils/i18nSettings";
 
 /**
  * These are the only genders that are apparent in the source data set,
