+++ src/components/charts/new_revocations/RevocationsByRiskLevel/helpers.js
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { translate } from "../../../../views/tenant_copy_translations/i18nSettings";
+import { translate } from "../../../../views/tenants/utils/i18nSettings";
 
 export const findDenominatorKeyByMode = (mode) => {
   switch (mode) {
