+++ src/utils/charts/info.js
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-import { translate } from "../../views/tenant_copy_translations/i18nSettings";
+import { translate } from "../../views/tenants/utils/i18nSettings";
 
 // prettier-ignore
 const chartIdToInfo = () => {
