+++ src/views/tenants/utils/coreTenants.js
