+++ src/views/tenants/utils/us_pa_translations.js
